\name{ASCAT.sc-package}
\alias{ASCAT.sc-package}
\alias{ASCAT.sc}
\docType{package}
\title{
\packageTitle{ASCAT.sc}
}
\description{
\packageDescription{ASCAT.sc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ASCAT.sc}
\packageIndices{ASCAT.sc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ASCAT.sc}

Maintainer: \packageMaintainer{ASCAT.sc}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
